🎨 Move internal monitoring into its folder
